backend/requirements.txt
# FlaskMarket Enterprise - Backend Dependencies
# Production-ready e-commerce platform

# Core Flask
Flask==3.0.0
Werkzeug==3.0.1

# Database
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.22
#psycopg2-binary==2.9.9  # PostgreSQL adapter

# Authentication & Security
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
python-dotenv==1.0.0

# API & Serialization
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
marshmallow==3.20.1

# CORS & Rate Limiting
Flask-CORS==4.0.0
Flask-Limiter==3.5.0

# File Uploads
#Pillow==10.1.0

# Email
Flask-Mail==0.9.1

# Caching & Performance
Flask-Caching==2.1.0
redis==5.0.1

# Task Queue (for background jobs)
celery==5.3.4

# Validation
email-validator==2.1.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Production Server
gunicorn==21.2.0

# Utilities
python-dateutil==2.8.2
requests==2.31.0
